+++ doc/news.rst
@@ -8,6 +8,8 @@ News / Changelog
 
 (20XX-XX-XX)
 
+- Allow ``SHOW_INACTIVE_LISTS_DEFAULT`` setting to be configurable. (Closes #276)
+
 
 1.3.2
 =====